test: Add more infrastructure unit tests.
I had already created these tests (please note
that they are not yet 100% complete).

I will not be including additional tests/fixes
because I don't believe it is the primary objective
of these tests. However, for this kind of
development of new features, TDD and unit tests
must always be provided, as well as E2E tests.
load and performance tests will probably be
performed as well, but I think they are outside
the scope of this test.